refactor(cf): relocate cf user signal configs out of list framework
