+++ src/frontend/packages/cloud-foundry/src/features/cf/tabs/cf-organizations/cf-organization-spaces/tabs/cloud-foundry-space-users/cloud-foundry-space-users.component.spec.ts
@@ -11,7 +11,7 @@ import { generateCfBaseTestModulesNoShared } from '@test-framework/cloud-foundry
 import { CloudFoundrySpaceUsersComponent } from './cloud-foundry-space-users.component';
 import {
   CfUsersSignalConfigService,
-} from '../../../../../../../shared/components/list/list-types/user/cf-users-signal-config.service';
+} from '../../../../../../../shared/signal-list-configs/user/cf-users-signal-config.service';
 import { CloudFoundryEndpointService } from '../../../../../services/cloud-foundry-endpoint.service';
 import { CloudFoundryOrganizationService } from '../../../../../services/cloud-foundry-organization.service';
 import { CloudFoundrySpaceService } from '../../../../../services/cloud-foundry-space.service';
