+++ src/frontend/packages/cloud-foundry/src/features/cf/tabs/cf-organizations/cf-organization-users/cloud-foundry-organization-users.component.ts
@@ -9,7 +9,7 @@ import {
   SignalListRowAction,
 } from '@stratosui/core';
 
-import { CfUsersSignalConfigService } from '../../../../../shared/components/list/list-types/user/cf-users-signal-config.service';
+import { CfUsersSignalConfigService } from '../../../../../shared/signal-list-configs/user/cf-users-signal-config.service';
 import { CloudFoundryEndpointService } from '../../../services/cloud-foundry-endpoint.service';
 import { CloudFoundryOrganizationService } from '../../../services/cloud-foundry-organization.service';
 import type { StUser, StUserOrgRole, StUserSpaceRole } from '../../../../../services/endpoint-data/stratos-types';
